Cast to signed int

+++ vlasovsolver/cpu_acc_map.cpp
@@ -320,10 +320,10 @@ bool map_1d(SpatialCell* spatial_cell,
          firstBlockIndexK = (firstBlockIndexK < max_v_length ) ? firstBlockIndexK : max_v_length - 1;
          lastBlockIndexK  = (lastBlockIndexK  >= 0)            ? lastBlockIndexK  : 0;
          lastBlockIndexK  = (lastBlockIndexK  < max_v_length ) ? lastBlockIndexK  : max_v_length - 1;
-         if(firstBlockIndexK < Parameters::bailout_velocity_space_wall_margin
-            || firstBlockIndexK >= max_v_length - Parameters::bailout_velocity_space_wall_margin
-            || lastBlockIndexK < Parameters::bailout_velocity_space_wall_margin
-            || lastBlockIndexK >= max_v_length - Parameters::bailout_velocity_space_wall_margin
+         if(firstBlockIndexK < static_cast<int>(Parameters::bailout_velocity_space_wall_margin)
+            || firstBlockIndexK >= max_v_length - static_cast<int>(Parameters::bailout_velocity_space_wall_margin)
+            || lastBlockIndexK < static_cast<int>(Parameters::bailout_velocity_space_wall_margin)
+            || lastBlockIndexK >= max_v_length - static_cast<int>(Parameters::bailout_velocity_space_wall_margin)
          ) {
             string message = "Some target blocks in acceleration are going to be less than ";
             message += std::to_string(Parameters::bailout_velocity_space_wall_margin);
